requirements.txt
numpy
scikit-learn
pyserial
matplotlib
openbci-python
joblib
scipy
pygame
